feat: Event Type Filter for Events History
